apps/automation_crews/requirements.txt
undetected-chromedriver # Chrome version should be 127 or greater.
pyautogui
PyVirtualDisplay
xvfbwrapper
pdfminer
pdfminer.six
python-dotenv
python-xlib
openai
discord.py
requests
opencv-python
webdriver_manager
setuptools
Pillow
webdriver-manager
selenium
crewai
hnswlib
crewai_tools
watchdog
composio-crewai
langchain-openai
groq
langchain-groq
flask
flask-cors
asgiref
google-generativeai 
python-dotenv
openai
gunicorn